requirements.txt
google-cloud-storage
google-auth
python-dotenv
faker
colorlog